[skip ci] - metric command working fine

+++ cloud/scripts/002-01-darcano.sh
@@ -37,6 +37,16 @@ helper.save_relay() {
     sudo systemctl restart cardano-node.service
 }
 
+helper.metrics() {
+	local metric=\$1
+	arr=(\${metric})
+	arr[0]="/metric"
+	metric=(\${arr[@]:1})
+
+	message=\$(curl 127.0.0.1:12798/metrics | grep -i \${metric} | awk '{ print \$2}')
+    ShellBot.sendMessage --chat_id \${message_chat_id[$id]} --text "\$(echo -e \${message})" --parse_mode markdown
+}
+
 helper.get_api
 
 source \${tmp_folder}/ShellBot.sh
@@ -51,19 +61,24 @@ do
 	for id in \$(ShellBot.ListUpdates)
 	do
 	(
-        if [[ "\$(echo \${message_text[\$id]%%@*} | grep "^\/relay" )" ]]; then
-		    helper.save_relay "\${message_text[\$id]}"
-            ShellBot.sendMessage --chat_id \${message_chat_id[\$id]} --text "done, bye" --parse_mode markdown
-        fi
-        if [[ "\$(echo \${message_text[\$id]%%@*} | grep "^\/txsProcessedNum" )" ]]; then
-		    message="Tx Processed: "
-            message+=\$(curl 127.0.0.1:12798/metrics | grep -i cardano_node_metrics_txsProcessedNum | awk '{ print \$2}')
-            ShellBot.sendMessage --chat_id \${message_chat_id[\$id]} --text "\$(echo -e \${message})" --parse_mode markdown
-        fi
-        if [[ "\$(echo \${message_text[\$id]%%@*} | grep "^\/kill-darcano" )" ]]; then
-            ShellBot.sendMessage --chat_id \${message_chat_id[\$id]} --text "done, bye" --parse_mode markdown
-            sleep 2
-            sudo systemctl stop darcano-bot
+        if [[ ${message_entities_type[$id]} == bot_command ]]; then
+            if [[ "\$(echo \${message_text[\$id]%%@*} | grep "^\/relay" )" ]]; then
+                helper.save_relay "\${message_text[\$id]}"
+                ShellBot.sendMessage --chat_id \${message_chat_id[\$id]} --text "done, bye" --parse_mode markdown
+            fi
+            if [[ "\$(echo \${message_text[\$id]%%@*} | grep "^\/txsProcessedNum" )" ]]; then
+                message="Tx Processed: "
+                message+=\$(curl 127.0.0.1:12798/metrics | grep -i cardano_node_metrics_txsProcessedNum | awk '{ print \$2}')
+                ShellBot.sendMessage --chat_id \${message_chat_id[\$id]} --text "\$(echo -e \${message})" --parse_mode markdown
+            fi
+            if [[ "\$(echo \${message_text[\$id]%%@*} | grep "^\/metric" )" ]]; then
+		        helper.metrics "\${message_text[\$id]}"
+            fi
+            if [[ "\$(echo \${message_text[\$id]%%@*} | grep "^\/kill-darcano" )" ]]; then
+                ShellBot.sendMessage --chat_id \${message_chat_id[\$id]} --text "done, bye" --parse_mode markdown
+                sleep 2
+                sudo systemctl stop darcano-bot
+            fi
         fi
 	) &
 	done
